requirements.txt
absl-py==0.9.0
alabaster==0.7.12
appnope==0.1.0
astor==0.8.1
atari-py==0.2.6
attrs==20.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.5
blis==0.4.1
boto==2.49.0
boto3==1.12.1
botocore==1.15.1
box2d-py==2.3.8
bz2file==0.98
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dlib==19.16.0
docutils==0.15.2
dopamine-rl==3.0.1
emnist==0.0
entrypoints==0.3
fire==0.2.1
Flask==1.1.2
future==0.18.2
gast==0.2.2
gensim==3.8.1
gevent==1.5.0
gin-config==0.3.0
glfw==1.10.1
google-api-python-client==1.8.0
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.27.1
gunicorn==20.0.4
gym==0.17.1
h5py==2.10.0
httplib2==0.17.2
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==1.5.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.1.0
lockfile==0.12.2
Markdown==3.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mesh-tensorflow==0.1.13
mistune==0.8.4
mpmath==1.1.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.16.1
numpydoc==0.9.1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==3.4.5.20
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pygame==2.0.0
pygifsicle==1.0.2
pyglet==1.4.10
Pygments==2.5.2
pygtrie==2.3.2
pyparsing==2.4.6
pypng==0.0.20
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
regex==2017.4.5
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
smart-open==1.9.0
snowballstemmer==2.0.0
sounddevice==0.3.14
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==1.0.1
stable-baselines==2.10.0
statsmodels==0.12.2
sympy==1.5.1
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-gpu==1.15.0
tensorflow-datasets==2.1.0
tensorflow-estimator==1.15.1
tensorflow-metadata==0.21.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.3.1
threadpoolctl==2.0.0
torch==1.5.0
torchfile==0.1.0
torchvision==0.6.0
tornado==6.0.4
tqdm==4.31.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.24.3
virtualenv==16.7.8
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.0.0
